feat: 카드의 칼럼이동 방식에 추가된 점 있음, comment 부분 유효성 검사를 위해 조금 수정함

+++ services/comments.services.js
@@ -12,9 +12,9 @@ class CommentsService {
         return { code: 200, message: '댓글 작성이 완료되었습니다.' };
     };
 
-    findOneComment = async ({ card_id }) => {
+    findAllComment = async ({ card_id }) => {
         try {
-            const allComments = await this.commentsRepository.findOneComment({ card_id });
+            const allComments = await this.commentsRepository.findAllComment({ card_id });
             const Comment = allComments.map((comment) => {
                 return {
                     comment_id: comment.comment_id,
@@ -30,24 +30,21 @@ class CommentsService {
     };
 
     updateComment = async (user_id, comment_id, content) => {
-        // const findComment = await this.commentsRepository.findOneComment(comment_id);
-        // if (!findComment) throw new Error('수정가능한 댓글이 존재하지 않습니다.');
+        const findComment = await this.commentsRepository.findOneComment(comment_id);
+        if (!findComment) throw new Error('수정가능한 댓글이 존재하지 않습니다.');
 
-        // const findCommentUser_id = findComment.user_id;
-        // if (user_id !== findCommentUser_id) throw new Error('댓글수정은 댓글작성자만 가능합니다.');
+        const findCommentUser_id = findComment.user_id;
+        if (user_id !== findCommentUser_id) throw new Error('댓글수정은 댓글작성자만 가능합니다.');
 
         await this.commentsRepository.updateComment(comment_id, content);
     };
 
     deleteComment = async (user_id, comment_id) => {
-        // const findCard = await this.cardRepository.cardFindOne(card_id);
-        // if (!findCard) throw { code: 403, message: '게시글이 존재하지 않습니다.' };
+        const findComment = await this.commentsRepository.findOneComment(comment_id);
+        if (!findComment) throw new Error('수정가능한 댓글이 존재하지 않습니다.');
 
-        // const findComment = await this.commentsRepository.findOneComment(comment_id);
-        // if (!findComment) throw { code: 403, message: '삭제가능한 댓글이 존재하지 않습니다.' };
-
-        // const findCommentUser_id = findComment.User_id;
-        // if (user_id !== findCommentUser_id) throw { code: 403, message: '댓글삭제는 댓글작성자만 가능합니다.' };
+        const findCommentUser_id = findComment.user_id;
+        if (user_id !== findCommentUser_id) throw new Error('댓글수정은 댓글작성자만 가능합니다.');
 
         await this.commentsRepository.deleteComment(comment_id);
         return { code: 200, message: '댓글삭제가 완료되었습니다.' };
